Prevent parallel updates per switch
Since some switch models will throw errors when multiple sessions
attempt to modify and commit system state, this patch introduces
a distributed lock to prevent more than a single ansible process
from running at any given time for each switch. This may be
refined in subsequent patches to improve exactly when this behavior
is required.

In addition, because requests are now required to wait for a lock
on the switch(es) they are intending to modify, requests may be
satisfied out of order. This means that each request now needs to
refresh its view of the state of the object it is modifying upon
acquisition of the lock, and may in fact not need to do anything.

An example of the latter would be if a vlan is added, removed and
then added again in quick succession. No matter which order these
requests are handled in, the final result will always be that the
vlan is present, since that was the last request and what will be
present in the neutron db.

Change-Id: I2562a9dd3de8d3a396ccf1a46d8e8bf5ff9f2a92