Move to python3
On python2 based distros using yum to grab ansible
will result in the py2 version, which will then lead
to errors in ansible-runner since it won't be able
to find ansible in its python path.

As python3 is a requirement going forward, we can
simply install ansible using python3.

In addition, the dev environment has been moved to
python3

Change-Id: I0bd2fcc43cffa03dfe1d308168d231982b3a4daf