Correct port detachment
The db was being checked to see if a port was deleted, but if it
is only detached then it will still be in the db despite a port
delete operation being desired. This is fixed by checking to see
if there's local link information on the port to determine if its
attached to anything.

Change-Id: I3231a3bd71e28e3f7e5271ee7aaf39063c098252